[expo-av] Split AVModule exported module into internal and exported modules
